2981. Find Longest Special Substring That Occurs Thrice I.
