+++ src/main/java/com/coding/challenge/dp/FindLongestSpecialSubstringThatOccursThriceI.java
@@ -1,7 +1,5 @@
 package com.coding.challenge.dp;
 
-import java.util.Arrays;
-
 class FindLongestSpecialSubstringThatOccursThriceI {
 	private static final char FIRST_LETTER = 'a';
 
@@ -10,17 +8,10 @@ class FindLongestSpecialSubstringThatOccursThriceI {
 	}
 
 	public static void main(String[] args) {
-		assert maximumLengthLinear("aaaa") == 2;
 		assert maximumLength("aaaa") == 2;
-		
-		assert maximumLengthLinear("abcdef") == -1;
-		assert maximumLength("abcdef") == -1;
-		
+		assert maximumLength("abcdef") == -1;	
 		assert maximumLength("abcaba") == 1;
-		assert maximumLengthLinear("abcaba") == 1;
-		
 		assert maximumLength("cccerrrecdcdccedecdcccddeeeddcdcddedccdceeedccecde") == 2;
-		assert maximumLengthLinear("cccerrrecdcdccedecdcccddeeeddcdcddedccdceeedccecde") == 2;
 	}
 
 	static int maximumLength(String s) {
@@ -55,52 +46,5 @@ class FindLongestSpecialSubstringThatOccursThriceI {
 					len = l;
 		}
 		return len;
-	}
-
-	// string matching algorithm.
-	static int maximumLengthLinear(String s) {
-		final int n = s.length();
-		int len = -1;
-		final int[][] a = new int[26][3];
-		// init.
-		for (int i = 0; i < 26; i++) {
-			a[i][0] = -1;
-			a[i][1] = -1;
-			a[i][2] = -1;
-		}
-
-		for (int i = 0, c = 0; i < n; i++) {
-			c = c + 1;
-			if (i == n - 1 || s.charAt(i) != s.charAt(i + 1)) {
-				final int[] arr = a[s.charAt(i) - FIRST_LETTER];
-				final int j = min(arr[0], arr[1], arr[2]);
-				if (c > arr[j])
-					arr[j] = c;
-				c = 0;
-			}
-		}
-
-		for (int[] f : a) {
-			Arrays.sort(f);
-			len = Math.max(len, f[0]);
-			if (f[2] > 1)
-				len = Math.max(len, Math.min(f[1], f[2] - 1));
-			if (f[2] > 2)
-				len = Math.max(len, f[2] - 2);
-		}
-		return len;
-	}
-
-	private static int min(int a, int b, int c) {
-		if (a <= b) {
-			if (a <= c)
-				return 0;
-
-			return 2;
-		} else {
-			if (b <= c)
-				return 1;
-			return 2;
-		}
 	}
 }
